improvements to pipeline and plotting

+++ gctree_benchmark_direct_run.sh
@@ -4,7 +4,6 @@ set -eu
 source /home/wdumm/gctree-benchmark-new/gctreeenvpy3.10/bin/activate
 
 mkdir -p testdir
-rm -f testdir/*.p
 true_filepaths=testdir/truefilepathsmap.txt
 true_trees=testdir/truetreesmap.txt
 rm -f $true_filepaths
@@ -17,7 +16,7 @@ touch $true_trees
 # # This was the old glob pattern:
 # for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/*/obs-times-*/*/simu/selection/simu/event-*/; do
 
-for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-*/obs-times-(15|20|30|40|50)/simu/selection/simu/event-*/; do
+for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v6/seed-*/obs-times-(15|20|30|40|50)/simu/selection/simu/event-*/; do
     echo $gctreepath $(./get_true_tree_newick.py $gctreepath/simu_lineage_tree.p) >> $true_trees
 done
 # ## End building Newicks of true trees
@@ -25,9 +24,13 @@ done
 echo Done putting new trees in $true_trees
 
 count=0
+
+# # old glob pattern:
 # for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/*/obs-times-*/*/partis/gctree/iclust-*/; do
 
-for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-*/obs-times-(15|20|30|40|50)/partis/gctree/iclust-*/; do
+# ### Do the evaluation:
+rm -f testdir/*.p
+for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v6/seed-*/obs-times-(15|20|30|40|50)/partis/gctree/iclust-*/; do
     python gctree_benchmark_direct.py testdir/$count.p $gctreepath/gctree.out.inference.parsimony_forest.p HS5F_Mutability.csv HS5F_Substitution.csv $gctreepath/input-seqs.fa $gctreepath/meta.yaml $true_trees $gctreepath/outfile $gctreepath/abundances.csv XnaiveX
 
     echo $gctreepath $count >> $true_filepaths
@@ -36,15 +39,29 @@ for gctreepath in /fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-
 
 done
 
+# Do plotting:
 python plot_criterion_comparison.py
 
-echo "Edit gctree_benchmark_direct_run.sh with path to best replicate (paths are in" $true_filepaths") then rerun tree_scatter.py"
+echo "Building example tree scatters for all trees in example_sims.txt. To try more from the table above, you can look up their filepaths in " $true_filepaths " then rerun tree_scatter.py"
 
-gctreepath=/fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-0/obs-times-40/partis/gctree/iclust-46/
-gctreepath=/fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-0/obs-times-30/partis/gctree/iclust-20/
-gctreepath=/fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-0/obs-times-30/partis/gctree/iclust-26/
-gctreepath=/fh/fast/matsen_e/dralph/partis/paired-loci/gct-valid/v4/seed-0/obs-times-50/partis/gctree/iclust-0/
-python gctree_benchmark_direct.py ignorethisfile.p $gctreepath/gctree.out.inference.parsimony_forest.p HS5F_Mutability.csv HS5F_Substitution.csv $gctreepath/input-seqs.fa $gctreepath/meta.yaml $true_trees $gctreepath/outfile $gctreepath/abundances.csv XnaiveX -a testdir/all_dagtrees_example.p
 
-rm ignorethisfile.p
-python tree_scatter.py
+rm -rf example_sim_scatters
+mkdir example_sim_scatters
+
+cat example_sims.txt | while read -r sim; do
+    # For each simulation number, find the corresponding line in truefilepathsmap.txt
+    gctreepath=$(grep " $sim$" $true_filepaths | awk '{print $1}' | head -n 1)
+    
+    # Check if a match was found
+    if [ -n "$gctreepath" ]; then
+        echo "Simulation $sim: $gctreepath"
+        python gctree_benchmark_direct.py ignorethisfile.p $gctreepath/gctree.out.inference.parsimony_forest.p HS5F_Mutability.csv HS5F_Substitution.csv $gctreepath/input-seqs.fa $gctreepath/meta.yaml $true_trees $gctreepath/outfile $gctreepath/abundances.csv XnaiveX -a testdir/all_dagtrees_example.p
+
+        rm ignorethisfile.p
+        python tree_scatter.py
+        mv tree_scatter.pdf example_sim_scatters/$sim.pdf
+
+    else
+        echo "Simulation $sim: No matching file path found"
+    fi
+done
